Refactors the multiple return solr queries.
